+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 ### Added
 
 * nostr: add `SingleLetterTag::as_str` and `TagKind::as_str` ([Yuki Kishimoto])
+* nostr: add `Kind::Comment` ([reyamir])
 
 ### Fixed